# SM75 (Turing) support for FP6 kernel