@gau-nernst 2024-09-25
@tobiasvanderwerff Thank you for your contribution! The debug when m>=64 is very insightful.

I was the one added this FP6-LLM kernel to torchao. I don't have enough CUDA knowledge to fully understand the changes, but it seems reasonable to me. Have you had the chance to discuss this with the original author about this change? https://github.com/usyd-fsalab/fp6_llm

I will let @msaroufim decide if we should support / improve the kernel for sm75 at all. Concerns include CI testing + deviation from upstream, as well as not worth the efforts due to old age.

General comments:
- Low performance for large M is to be expected. I don't think you need to try to optimize for it
- Regarding guard for M>=64, if the macro doesn't work, I think you can check with either `torch.cuda.get_device_capability()` or `cudaDeviceProp`
- Regarding the splitK, yea I don't really like it too. I wish we have an autotune for CUDA kernel, similar to triton 😄
- (Can do later, once everything is finalized) Can you add a summary of your changes at the top of each modified file?

Side question. Would you be able to modify this kernel to support BF16? Since all modern models use BF16, it would be useful. I figure it would be a matter of changing the dequant logic + MMA instructions, but again, I'm not too confident with my CUDA skills 😅.