@gau-nernst 2024-09-26
@tobiasvanderwerff Thank you for your response

I'm just checking if you ask the original author about this change since he probably can give more feedback in terms of accuracy and perf wise. No worries!

Since bad perf for large M is to be expected (actually FP6-LLM kernel is also slower than FP16 on A100 for M>=128), I think we don't need to check for M>=64? Just use `__CUDA_ARCH__` to select tile size accordingly.

> Err, you mean a comment in the C++/CUDA files? As in, add a comment at the top of each changed file saying how I added SM75 support to it? If so, I could do definitely do that.

Yes, exactly.

> I'd love to try that, but unfortunately the T4 does not support BF16 (it requires SM80). So that would require an A100 to work with, which is currently not feasible for me.

No worries at all! In case you don't know, for sm80 stuff, Lightning AI gives out free 22hrs/month of L4. Some GPU clouds also provide rental service for consumer cards, like 3090 and 4090, which is much cheaper than A100. Just something to consider.

Some other pointers
- I think the build wheel CI does not compile for sm75. So interested users have to compile from source. Can you add a line in `torchao/dtypes/floatx/README.md` saying that sm75 support requires installing torchao from source?
https://github.com/pytorch/ao/blob/64719d5882ac3790e0018bed77d2d7b1e788ac5c/packaging/env_var_script_linux.sh#L16-L19
- If possible, can you run the llama2 benchmark for sanity check? https://github.com/pytorch/ao/blob/main/torchao/dtypes/floatx/README.md#end-to-end-benchmarks