@tobiasvanderwerff 2024-09-26
The kernel now handles the `m>=64` edge case to produce correct results for all shapes used in the benchmark (see https://github.com/pytorch/ao/pull/942/commits/650ba03d7df2bbb5c78e0620f71911ecc6c60927). 

I couldn't use the __CUDA_ARCH__ constant, so I'm instead checking the CUDA arch at runtime. It unfortunately makes the kernel launch code a bit uglier, since the template parameters of `TilingConfig` need to be known at compile time. I'm not the most experienced with C++ so there may be a better solution, but this is the best I could come up with to handle this.