@tobiasvanderwerff 2024-09-27
# Benchmark results on Llama-2-7b-chat-hf

Testing specs:
- T4 GPU
- Torch version: 2.4.1

## float16

```bash
python generate.py --compile --precision float16
```

Average tokens/sec: 19.04
Average Bandwidth: 251.60 GB/s
Peak Memory Usage: 13.90 GB
Model Size: 13.21 GB



## fp6

```bash
python generate.py --compile --precision float16 --quantization fp6
```

Average tokens/sec: 40.52
Average Bandwidth: 200.93 GB/s
Peak Memory Usage: 6.61 GB
Model Size: 4.96 GB