@tobiasvanderwerff 2024-09-27
# Fp6 eval results

```bash
python eval.py --compile --precision float16 --quantization fp6
```

wikitext:
- word_perplexity: 12.3653
- byte_perplexity: 1.6005
- bits_per_byte: 0.6785